refactor: 파일명 변경 및 로컬 데이터베이스 모듈 생성
