+++ app/src/main/java/com/wasingun/seller_lounge/local/RecentPostDao.kt
@@ -9,7 +9,7 @@ import com.wasingun.seller_lounge.data.model.localpost.LocalPostInfo
 import com.wasingun.seller_lounge.data.model.post.PostInfo
 
 @Dao
-interface LocalPostDao {
+interface RecentPostDao {
 
     @Query("SELECT * FROM LocalPostInfo ORDER BY savedTime DESC")
     suspend fun getPostInfo(): List<LocalPostInfo>
